Added a new unit test to check whether or not aaply files when called across an array list (as described in the JStatSoft paper regarding the 'ozone' dataset).
Fixed the list_to_array code called by laply to correctly check
whether or not the 'res' object (from llply) is from an array list,
and deal with it appropriately.

Checked that the new tests worked with the modified code, and that
no other tests failed as a result of the new code.