Keep .progress="progressr" even when .parallel=TRUE (#301)
This PR makes **dplyr** report on progress also when `.parallel = TRUE`. For this to work, one need to use **doFuture** for parallel processing and **progressr** for progress updates.  The problem with **plyr** 1.8.6 is that it forces `.progress = "none" whenever `.parallel = TRUE`.

Example of `progress::progress_bar()` updates when running in parallel:

```r
>library(plyr)
> library(progressr)
> doFuture::registerDoFuture()
> future::plan("multisession", workers = 2L)
> handlers("progress")  ## 'progress' package

> with_progress({
+   y <- llply(1:10, function(x) { Sys.sleep(1); identity(x) },
+                 .parallel = TRUE, .progress = "progressr")
+ })
[================>---------------------------------------]  20%
```